+++ Packages/UGF.Types/Editor/TypesEditorUtility.cs
@@ -1,0 +1,9 @@
+﻿namespace UGF.Types.Editor
+{
+    /// <summary>
+    /// Provides utilities to work with types in editor.
+    /// </summary>
+    public static class TypesEditorUtility
+    {
+    }
+}