+++ Packages/UGF.Types/Runtime/TypesUtility.cs
@@ -75,6 +75,53 @@ namespace UGF.Types.Runtime
         }
 
         /// <summary>
+        /// Collects all types into the specified collection that match by specified func condition, if presents.
+        /// <para>
+        /// If validate func does not specified, will add all types.
+        /// </para>
+        /// </summary>
+        /// <param name="types">The collection to add types.</param>
+        /// <param name="validate">The function to validate type.</param>
+        public static void CollectTypes(ICollection<Type> types, Func<Type, bool> validate = null)
+        {
+            if (types == null) throw new ArgumentNullException(nameof(types));
+
+            Assembly[] assemblies = AppDomain.CurrentDomain.GetAssemblies();
+
+            for (int i = 0; i < assemblies.Length; i++)
+            {
+                CollectTypes(types, assemblies[i], validate);
+            }
+        }
+
+        /// <summary>
+        /// Collects all types the specified assembly into the specified collection, that match by specified func condition, if presents.
+        /// <para>
+        /// If validate func does not specified, will add all types from the assembly.
+        /// </para>
+        /// </summary>
+        /// <param name="types">The collection to add types.</param>
+        /// <param name="assembly">The assembly to gather types.</param>
+        /// <param name="validate">The function to validate type.</param>
+        public static void CollectTypes(ICollection<Type> types, Assembly assembly, Func<Type, bool> validate = null)
+        {
+            if (types == null) throw new ArgumentNullException(nameof(types));
+            if (assembly == null) throw new ArgumentNullException(nameof(types));
+
+            Type[] assemblyTypes = assembly.GetTypes();
+
+            for (int i = 0; i < assemblyTypes.Length; i++)
+            {
+                Type type = assemblyTypes[i];
+
+                if (validate == null || validate(type))
+                {
+                    types.Add(type);
+                }
+            }
+        }
+
+        /// <summary>
         /// Tries to get type identifier from the specified type that contains identifier attribute.
         /// </summary>
         /// <param name="type">The target type.</param>